hardware/ml_models/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
blinker==1.9.0
builtin-interfaces==1.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
composition-interfaces==1.2.1
contourpy==1.3.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.10.0
geometry-msgs==4.2.4
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.2
matplotlib==3.9.2
message-filters==4.3.5
mpmath==1.3.0
nav-msgs==4.2.4
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
osrf-pycommon==2.0.2
packaging==24.2
pandas==2.2.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pi_img_to_cv2==0.0.0
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.2
PyYAML==6.0.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
requests==2.32.3
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.14.1
seaborn==0.13.2
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.1
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
topic-monitor==0.20.5
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
trajectory-msgs==4.2.4
tree_detection==0.0.0
triton==3.1.0
turtlesim==1.4.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
unique-identifier-msgs==2.2.1
urllib3==2.2.3
visualization-msgs==4.2.4
Werkzeug==3.1.3
